Extracting the related content wrapper up
For parity between related content and primary content I'm doing this.
I have usually found with partials there is the concept of the container
and contents. I don't want to render the container with the contents.

In doing this, we preserve the dom classes and thus can be "more
certain" that our styles are applied.